Front/avatar e integração api

+++ backend/src/main/java/com/ft/trans/service/GamificationService.java
@@ -107,7 +107,7 @@ public class GamificationService
             }
         }
 
-        Integer totalXp = safeTotalXp(user.id);
+        Long totalXp = safeTotalXp(user.id);
         Level currentLevel = levelRepository.findTopByXpRequiredLessThanEqualOrderByXpRequiredDesc(totalXp);
         Level nextLevel = levelRepository.findTopByXpRequiredGreaterThanOrderByXpRequiredAsc(totalXp);
 
@@ -134,7 +134,7 @@ public class GamificationService
         if (user == null)
             return new SummaryResult(false, "Usuario nao encontrado.", null);
 
-        Integer totalXp = safeTotalXp(userId);
+        Long totalXp = safeTotalXp(userId);
         Level currentLevel = levelRepository.findTopByXpRequiredLessThanEqualOrderByXpRequiredDesc(totalXp);
         Level nextLevel = levelRepository.findTopByXpRequiredGreaterThanOrderByXpRequiredAsc(totalXp);
 
@@ -165,9 +165,9 @@ public class GamificationService
         return new SummaryResult(true, "Resumo carregado com sucesso.", response);
     }
 
-    private Integer safeTotalXp(Long userId)
+    private Long safeTotalXp(Long userId)
     {
-        Integer total = xpHistoryRepository.sumXpByUserId(userId);
+        Long total = xpHistoryRepository.sumXpByUserId(userId);
         return total != null ? total : 0;
     }
